requirements.txt
pdfplumber
PyYAML
